Small bug fix and small update in function and pagination add

+++ routes/EmployeeManagement/employee.routes.php
@@ -5,9 +5,9 @@ use App\Http\Controllers\EmployeeController;
 
 Route::middleware(['web', 'auth', 'permission:view_employees'])->group(function () {
     Route::get('/employees', [EmployeeController::class, 'index'])->name('employees.index');
-    Route::get('/employees/create', [EmployeeController::class, 'create'])->name('employees.create');
-    Route::post('/employees', [EmployeeController::class, 'store'])->name('employees.store');
+    Route::get('/employees/create', [EmployeeController::class, 'create'])->name('employees.create')->middleware('permission:manage_employees');
+    Route::post('/employees', [EmployeeController::class, 'store'])->name('employees.store')->middleware('permission:manage_employees');
     Route::get('/employees/{employee}/edit', [EmployeeController::class, 'edit'])->name('employees.edit');
     Route::put('/employees/{employee}', [EmployeeController::class, 'update'])->name('employees.update');
-    Route::delete('/employees/{employee}', [EmployeeController::class, 'destroy'])->name('employees.destroy');
+    Route::delete('/employees/{employee}', [EmployeeController::class, 'destroy'])->name('employees.destroy')->middleware('permission:manage_employees');
 });
